v0.74
This version brings along an entirely different way to gather your purchase data.  By using Humble Bundle's API much more information can be gathered for the user.
I also added in some more options to filter the data retrieved from the API.

If API Call is not enabled in the options the filters for it will be hidden , also if retrieve all API data is enabled the filters will be hidden.
Fetching choices (monthlies) is currently only for the scraper.  It will be hidden if API Call is enabled.